+++ fesod-sheet/src/test/java/org/apache/fesod/sheet/annotation/AnnotationMetadataTest.java
@@ -27,12 +27,15 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import org.apache.fesod.sheet.testkit.Tags;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
 /**
  * Tests for {@link AnnotationMetadata}
  */
+@Tag(Tags.UNIT)
 class AnnotationMetadataTest {
 
     @Target({ElementType.FIELD})
