+++ fesod-sheet/src/test/java/org/apache/fesod/sheet/annotation/composite/DirectAnnotationTest.java
@@ -45,9 +45,11 @@ import org.apache.fesod.sheet.metadata.ConfigurationHolder;
 import org.apache.fesod.sheet.metadata.GlobalConfiguration;
 import org.apache.fesod.sheet.metadata.Head;
 import org.apache.fesod.sheet.metadata.property.ExcelHeadProperty;
+import org.apache.fesod.sheet.testkit.Tags;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
@@ -82,6 +84,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
  *       per-field independence with shared class-level annotations</li>
  * </ul>
  */
+@Tag(Tags.UNIT)
 @ExtendWith(MockitoExtension.class)
 class DirectAnnotationTest {
